+++ packages/react-ui/src/widgets/utils/getChartSerie.js
@@ -1,7 +1,0 @@
-
-export default function getChartSerie (chart, index) {
-  const option = chart.getOption();
-  const serie =  option.series[index];
-
-  return { option, serie };
-};